PostgreSql Dialect Refactory
