+++ Sharp.Data/Sharp.Data.csproj
@@ -75,9 +75,13 @@
     <Compile Include="Databases\DatabaseKind.cs" />
     <Compile Include="Databases\Oracle\OracleReflectionCache.cs" />
     <Compile Include="Databases\PostgreSql\PostgreDbFactory.cs" />
+    <Compile Include="Databases\PostgreSql\PostgreSqlColumnDialect.cs" />
+    <Compile Include="Databases\PostgreSql\PostgreSqlConstraintsDialect.cs" />
     <Compile Include="Databases\PostgreSql\PostgreSqlDataClient.cs" />
+    <Compile Include="Databases\PostgreSql\PostgreSqlDbTypesDialect.cs" />
     <Compile Include="Databases\PostgreSql\PostgreSqlDialect.cs" />
     <Compile Include="Databases\PostgreSql\PostgreSqlProvider.cs" />
+    <Compile Include="Databases\PostgreSql\PostgreSqlTableDialect.cs" />
     <Compile Include="Databases\PostgreSql\SqlColumnStructure.cs" />
     <Compile Include="DbFactory.cs" />
     <Compile Include="DefaultDatabase.cs" />
